+++ packages/pros/examples/accessories.rs
@@ -7,8 +7,9 @@ use alloc::sync::Arc;
 use core::time::Duration;
 
 use pros::{
+    color::Rgb,
     devices::{
-        smart::vision::{LedMode, Rgb, VisionZeroPoint},
+        smart::vision::{LedMode, VisionZeroPoint},
         Controller,
     },
     prelude::*,
@@ -50,8 +51,6 @@ impl AsyncRobot for ExampleRobot {
 
         self.vision.set_led(LedMode::On(Rgb::new(0, 0, 255)));
 
-        pros::lcd::buttons::register(left_button_callback, Button::Left);
-
         // Spawn a new task that will print whether or not the motor is stopped constantly.
         spawn({
             let motor = Arc::clone(&self.motor); // Obtain a shared reference to our motor to safely share between tasks.
@@ -75,7 +74,7 @@ impl AsyncRobot for ExampleRobot {
             // Set output takes a float from -1 to 1 that is scaled to -12 to 12 volts.
             self.motor
                 .lock()
-                .set_output(controller.state().joysticks.right.y)?;
+                .set_output(controller.state()?.joysticks.right.y)?;
 
             // println!("pid out {}", pid.update(10.0, motor.position().into_degrees() as f32));
             println!(
@@ -92,7 +91,3 @@ async_robot!(
     ExampleRobot,
     ExampleRobot::new(Peripherals::take().unwrap())
 );
-
-fn left_button_callback() {
-    println!("Left button pressed!");
-}