+++ packages/pros/src/devices/adi/potentiometer.rs
@@ -1,9 +1,12 @@
-use pros_sys::{adi_potentiometer_type_e_t, ext_adi_potentiometer_t, PROS_ERR};
+//! ADI Potentiometer device.
+
+use pros_sys::{adi_potentiometer_type_e_t, ext_adi_potentiometer_t, PROS_ERR, PROS_ERR_F};
 
 use super::{AdiDevice, AdiDeviceType, AdiError, AdiPort};
 use crate::error::bail_on;
 
 #[derive(Debug, Eq, PartialEq)]
+/// Analog potentiometer ADI device.
 pub struct AdiPotentiometer {
     potentiometer_type: AdiPotentiometerType,
     raw: ext_adi_potentiometer_t,
@@ -28,27 +31,31 @@ impl AdiPotentiometer {
         })
     }
 
-    pub fn potentiometer_type(&self) -> AdiPotentiometerType {
+    /// Get the type of ADI potentiometer device.
+    pub const fn potentiometer_type(&self) -> AdiPotentiometerType {
         self.potentiometer_type
     }
 
-    /// Gets the current potentiometer angle in tenths of a degree.
+    /// Gets the current potentiometer angle in degrees.
     ///
     /// The original potentiometer rotates 250 degrees
     /// thus returning an angle between 0-250 degrees.
     /// Potentiometer V2 rotates 330 degrees
     /// thus returning an angle between 0-330 degrees.
     pub fn angle(&self) -> Result<f64, AdiError> {
-        Ok(bail_on!(PROS_ERR.into(), unsafe {
+        Ok(bail_on!(PROS_ERR_F, unsafe {
             pros_sys::ext_adi_potentiometer_get_angle(self.raw)
-        }))
+        }) / 10.0)
     }
 }
 
 #[derive(Debug, Clone, Copy, Eq, PartialEq)]
 #[repr(i32)]
+/// The type of potentiometer device.
 pub enum AdiPotentiometerType {
+    /// EDR potentiometer.
     PotentiometerEdr = pros_sys::E_ADI_POT_EDR,
+    /// V2 potentiometer.
     PotentiometerV2 = pros_sys::E_ADI_POT_V2,
 }