+++ packages/pros/src/devices/smart/motor.rs
@@ -52,6 +52,7 @@ impl Motor {
         Ok(motor)
     }
 
+    /// Sets the gearset of the motor.
     pub fn set_gearset(&mut self, gearset: Gearset) -> Result<(), MotorError> {
         bail_on!(PROS_ERR, unsafe {
             pros_sys::motor_set_gearing(self.port.index(), gearset as i32)
@@ -59,6 +60,7 @@ impl Motor {
         Ok(())
     }
 
+    /// Gets the gearset of the motor.
     pub fn gearset(&self) -> Result<Gearset, MotorError> {
         unsafe { pros_sys::motor_get_gearing(self.port.index()).try_into() }
     }
@@ -219,7 +221,7 @@ impl Motor {
     }
 
     /// Returns a future that completes when the motor reports that it has stopped.
-    pub fn wait_until_stopped(&self) -> MotorStoppedFuture {
+    pub const fn wait_until_stopped(&self) -> MotorStoppedFuture<'_> {
         MotorStoppedFuture { motor: self }
     }
 }
@@ -270,7 +272,9 @@ impl From<BrakeMode> for pros_sys::motor_brake_mode_e_t {
 /// Represents what the physical motor is currently doing.
 #[derive(Debug, Clone, Default)]
 pub struct MotorState {
+    /// The motor is currently moving.
     pub busy: bool,
+    /// the motor is not moving.
     pub stopped: bool,
     /// the motor is at zero encoder units of rotation.
     pub zeroed: bool,
@@ -291,8 +295,11 @@ impl From<u32> for MotorState {
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 #[repr(i32)]
 pub enum Gearset {
+    /// 36:1 gear ratio
     Red = pros_sys::E_MOTOR_GEAR_RED,
+    /// 18:1 gear ratio
     Green = pros_sys::E_MOTOR_GEAR_GREEN,
+    /// 6:1 gear ratio
     Blue = pros_sys::E_MOTOR_GEAR_BLUE,
 }
 
@@ -333,6 +340,9 @@ impl TryFrom<pros_sys::motor_gearset_e_t> for Gearset {
     }
 }
 
+#[derive(Debug)]
+/// A future that completes when the motor reports that it has stopped.
+/// Created by [`Motor::wait_until_stopped`]
 pub struct MotorStoppedFuture<'a> {
     motor: &'a Motor,
 }
@@ -354,11 +364,16 @@ impl<'a> core::future::Future for MotorStoppedFuture<'a> {
 }
 
 #[derive(Debug, Snafu)]
+/// Errors that can occur when using a motor.
 pub enum MotorError {
-    #[snafu(display("The voltage supplied was outside of the allowed range (-12 to 12)."))]
+    /// The voltage supplied was outside of the allowed range of [-12, 12].
     VoltageOutOfRange,
     #[snafu(display("{source}"), context(false))]
-    Port { source: PortError },
+    /// Generic port related error.
+    Port {
+        /// The source of the error.
+        source: PortError,
+    },
 }
 
 map_errno! {