+++ packages/pros/src/devices/smart/vision.rs
@@ -10,8 +10,8 @@ use snafu::Snafu;
 
 use super::{SmartDevice, SmartDeviceType, SmartPort};
 use crate::{
+    color::Rgb,
     error::{bail_errno, bail_on, map_errno, PortError},
-    lvgl::colors::LcdColor,
 };
 
 /// Represents a vision sensor plugged into the vex.
@@ -139,13 +139,20 @@ impl SmartDevice for VisionSensor {
 
 //TODO: figure out how coordinates are done.
 #[derive(Default, Debug, Clone, Copy, PartialEq, Eq)]
+/// An object detected by the vision sensor
 pub struct VisionObject {
+    /// The offset from the top of the object to the vision center.
     pub top: i16,
+    /// The offset from the left of the object to the vision center.
     pub left: i16,
+    /// The x-coordinate of the middle of the object relative to the vision center.
     pub middle_x: i16,
+    /// The y-coordinate of the middle of the object relative to the vision center.
     pub middle_y: i16,
 
+    /// The width of the object.
     pub width: i16,
+    /// The height of the object.
     pub height: i16,
 }
 
@@ -168,89 +175,50 @@ impl TryFrom<pros_sys::vision_object_s_t> for VisionObject {
     }
 }
 
-#[derive(Default, Debug, Clone, Copy, PartialEq, Eq)]
-pub struct Rgb {
-    r: u8,
-    g: u8,
-    b: u8,
-}
-
-impl Rgb {
-    pub fn new(r: u8, g: u8, b: u8) -> Self {
-        Self { r, g, b }
-    }
-}
-
-impl From<Rgb> for u32 {
-    fn from(other: Rgb) -> u32 {
-        ((other.r as u32) << 16) + ((other.g as u32) << 8) + other.b as u32
-    }
-}
-
-const BITMASK: u32 = 0b11111111;
-
-impl From<u32> for Rgb {
-    fn from(value: u32) -> Self {
-        Self {
-            r: ((value >> 16) & BITMASK) as _,
-            g: ((value >> 8) & BITMASK) as _,
-            b: (value & BITMASK) as _,
-        }
-    }
-}
-
-impl From<Rgb> for LcdColor {
-    fn from(other: Rgb) -> Self {
-        Self(pros_sys::lv_color_t {
-            red: other.r,
-            green: other.g,
-            blue: other.b,
-            alpha: 0xFF,
-        })
-    }
-}
-
-impl From<LcdColor> for Rgb {
-    fn from(other: LcdColor) -> Self {
-        Self {
-            r: other.red,
-            g: other.green,
-            b: other.blue,
-        }
-    }
-}
-
 #[repr(u32)]
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+/// The zero point of the vision sensor.
+/// Vision object coordinates are relative to this point.
 pub enum VisionZeroPoint {
+    /// The zero point will be the top left corner of the vision sensor.
     TopLeft,
+    /// The zero point will be the top right corner of the vision sensor.
     Center,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+/// The white balance of the vision sensor.
 pub enum WhiteBalance {
+    /// Provide a specific color to balance the white balance.
     Rgb(Rgb),
+    /// Automatically balance the white balance.
     Auto,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+/// The mode of the vision sensor led.
 pub enum LedMode {
+    /// Turn on the led with a certain color.
     On(Rgb),
+    /// Turn off the led.
     Off,
 }
 
 #[derive(Debug, Snafu)]
+/// Errors that can occur when using a vision sensor.
 pub enum VisionError {
-    #[snafu(display(
-        "The index specified was higher than the total number of objects seen by the camera."
-    ))]
+    /// The camera could not be read.
     ReadingFailed,
-    #[snafu(display("The camera could not be read."))]
+    /// The index specified was higher than the total number of objects seen by the camera.
     IndexTooHigh,
-    #[snafu(display("Port already taken."))]
+    /// Port already taken.
     PortTaken,
     #[snafu(display("{source}"), context(false))]
-    Port { source: PortError },
+    /// Generic port related error.
+    Port {
+        /// The source of the error.
+        source: PortError,
+    },
 }
 
 map_errno! {