+++ packages/pros/src/io.rs
@@ -1,4 +1,0 @@
-//! Helpers for terminal I/O functionality.
-
-pub use libc_print::std_name::*;
-pub use no_std_io::io::*;