+++ CHANGELOG.md
@@ -22,6 +22,57 @@ Before releasing:
 ## [Unreleased]
 
 ### Added
+- Added feedforward motor controllers (#80)
+- Lightly document all APIs with missing documentation. (#70)
+- Added `Debug`, `Copy`, and `Clone` derives for common structs (#70)
+
+- Screen drawing API. (#81)
+- Added screen field to `Peripherals` and `DynamicPeripherals::take_screen` method. (#81)
+- Added `AdiSolenoid`, a wrapper over `AdiDigitalOut` for actuating SMC pneumatic solenoids. (#61)
+- Added `AdiSwitch`, another `AdiDigitalOut` wrapper that abstracts bumper switches and limit switches. (#61)
+- Added `AdiLineTracker` for abstracting the EDR line tracker sensor.
+- Implements TryFrom for Gearset.
+- Adds support for getting brake modes from motors. (#66)
+
+### Fixed
+
+- Fix error handling and error type variats in ADI bindings
+
+### Changed
+
+- Re-exported printing macros from `pros::io`. (#82)
+- Applied several lints to improve code quality. (#70)
+- Updated to PROS version 4. (**Breaking Change**) (#81)
+- Moved `vision::Rgb` into its own `color.rs` file. (**Breaking Change**) (#81)
+- The VEXOS target has been updated to improve file size and floating point operation speed. (#81)
+- `Peripherals::new()` is no longer const (**Breaking Change) (#81)
+- Updated panic handler to print to the brain display as well as over serial (#81)
+- Refactors digital and analog ADI input/output. (**Breaking Change**) (#61)
+	- Adds LogicLevel rather than bools for controlling digital devices.
+	- Adds 0-5V voltage getters and setters for analog ADI.
+	- Changed analog getters and setters to use `u16` data.
+- Changed `AdiPotentiometer` to return degrees rather than tenth degrees (**Breaking Change**) (#61).
+	- Renamed `AdiPotentiometer::value` to `AdiPotentiometer::angle`.
+- Refactors `AdiMotor` to match the smart motor APIs, having output/raw output getters/setters.
+- Renamed `AdiUltrasonic::value` to `AdiUltrasonic::distance` (**Breaking Change**) (#61).
+- Renamed `AdiEncoder::value` to `AdiEncoder::position` (**Breaking Change**) (#61).
+- Repurposed `AdiAnalogOut` as `AdiPwmOut` to correct match port output. (**Breaking Change**) (#90).
+- Adjusts constructor arguments for `Motor` to allow passing `Gearset` and `reversed` instead of `brake_mode` at construction. (**Breaking Change**) (#66)
+- Renamed `Motor::get_state` to `Motor::state`. (**Breaking Change**) (#66)
+- Changed `Motor::reversed` to return `Result<bool, _>`` rather than just `false` if `PROS_ERR` is returned. (**Breaking Change**) (#66)
+
+### Removed
+
+- LVGL bindings (pros-sys) and colors (pros). (**Breaking Change**) (#81)
+- LLEMU/lcd bindings. (**Breaking Change**) (#81)
+- Re-exported printing macros from `pros::io`. (#82)
+- Applied several lints to improve code quality. (#70)
+- Removed the confusingly named `write`, `ewrite`, `writeln`, and `ewriteln` macros. (**Breaking Change**) (#82)
+- Removed AdiDigitalIn::new_press, instead swapping it for AdiSwitch::was_pressed. (**Breaking Change**) (#61)
+
+## [0.7.0]
+
+### Added
 
 - `SmartPort` struct for device access. (#34)
 - `SmartDevice` trait for common functionality across smart port devices. (#34)
@@ -36,8 +87,7 @@ Before releasing:
 - All ADI device bindings (#55)
 - `LocalKey` now has `Cell`/`RefCell`-specific methods for setting and taking values. (#42)
 - `Peripherals` and `DynamicPeripherals` structs to ensure that you have only registered one device on a given smart or ADI port. (#53)
-- Implements TryFrom for Gearset.
-- Adds support for getting brake modes from motors. (#66)
+- Support for ADI Expander modules with `AdiExpander`. (#63)
 
 ### Fixed
 
@@ -45,20 +95,23 @@ Before releasing:
 - Fixed error handling in IMU sensor bindings. (#37)
 - Fixed errors in doctests and examples throughout the crate. (#37)
 - Fixed Missing ERRNO and ADI config variants in pros-sys (#55)
+- Fixed incorrect error handling with `InertialSensor::status`. (#65)
+- `Controller::status` now handles errors by returning `Result<ControllerStatus, ControllerError>`. (**Breaking Change**) (#74)
 
 ### Changed
 
 - Overhauled the `competition` module with more straightforward getters for competition state. (#38) (**Breaking Change**)
 - LLEMU-related macros have been prefixed with `llemu_` (e.g. `llemu_println`). (**Breaking Change**) (#30)
 - Added `Debug`, `Copy`, and `Clone` derives for common structs (#37)
-- Adjusts constructor arguments for `Motor` to allow passing `Gearset` and `reversed` instead of `brake_mode` at construction. (**Breaking Change**) (#66)
-- Renamed `Motor::get_state` to `Motor::state`. (**Breaking Change**) (#66)
-- Changed `Motor::reversed` to return `Result<bool, _>`` rather than just `false` if `PROS_ERR` is returned. (**Breaking Change**) (#66)
+- Renamed `InertialSensor::is_calibrating` to `InertialSensor::calibrating`. (**Breaking Change**) (#65)
+- Battery API functions now return `Result<_, BatteryError>`. (**Breaking Change**) (#62)
+- Renamed `battery::get_capacity` to `battery::capacity`, `battery::get_current` -> `battery::current`, `battery::get_temperature` -> `battery::temperature`, `battery::get_voltage` -> `battery::voltage`. (**Breaking Change**) (#62)
 
 ### Removed
 
 - Removed several broken bindings in `pros_sys` relating to competition state. (#38) (**Breaking Change**)
 - `LocalKey` no longer implements `set` for non-`Cell`/`RefCell` stored values. (**Breaking change**) (#42)
+- Removed the now-redundant `InertialStatus::error` function. (**Breaking Change**) (#65)
 
 ## [0.6.0] - 2024-01-14
 
@@ -119,7 +172,8 @@ Before releasing:
 
 ### Removed
 
-[unreleased]: https://github.com/pros-rs/pros-rs/compare/v0.6.0...HEAD
+[unreleased]: https://github.com/pros-rs/pros-rs/compare/v0.7.0...HEAD
 [0.4.0]: https://github.com/pros-rs/pros-rs/releases/tag/v0.4.0
 [0.5.0]: https://github.com/pros-rs/pros-rs/compare/v0.4.0...v0.5.0
 [0.6.0]: https://github.com/pros-rs/pros-rs/compare/v0.5.0...v0.6.0
+[0.7.0]: https://github.com/pros-rs/pros-rs/compare/v0.6.0...v0.7.0